requirements.txt
openai==0.28.1
python-dotenv==1.0.0
PyPDF2==3.0.1
tiktoken==0.5.2
chromadb==0.4.18
sentence-transformers==2.5.1
numpy==1.24.3 